Disable Package Watchdog explicit health check during test
There is a potential race condition in this test, due
to the fact that on a test device, the normal Watchdog behavior
can interfere with the test's assertion of requested packages.
By disabling Package Watchdog's explicit health check during the
test, this race condition can be avoided.

Test: atest ExplicitHealthCheckServiceImplTest
Bug: 146767850

Change-Id: I20aa42f6dab1bb34b04aa0f63aa5f7e98c7c371c
Exempt-From-Owner-Approval: Committer is owner in the main Watchdog directory