Fix ExtServices crash problem when getting shared preferences
It will cause crash in ExtServices when LRResolverRankerService gets
shared preferences under locked status, as those are stored in credential
encrypted storage.

This is fixed by checking that the user has unlocked the device before
getting shared preferences (see
https://developer.android.com/training/articles/direct-boot#notification).
and retrying the initialization if needed when LRResolverRankerService
methods are called.

Bug: 154421549
Test: atest ExtServicesUnitTests
Change-Id: Ic82e1f6a67b9de1c2a99540fd7928fb5b3b00719