Remove Application level directBootAware bit.
The application level bit will override the directBootAware bit for all
its component, which is causing bugs in WorkManager, part of
TextClassifierService. WorkManager defined a lot of directBootAware=False components in its manifest. They are all not respected because of this. It causes crashes because these WorkManager components are not expected to run before the user is unlocked.

We are setting directBootAware=True to all other existing components in the manifest. So this change should have no impact to them, i.e. The directBootAware bit of them remains to be true.

TESTED=verified the directBootAware bit for all components in ExtServices before&after the fix. See more in the bug.

BUG: 219999453
Change-Id: I48ecd8e47bf9d1f611b2d413a069dcbccbbe9920
(cherry picked from commit 4549e2d46307998fe2bc275bb3b08abe8688b2c8)
Merged-In:I48ecd8e47bf9d1f611b2d413a069dcbccbbe9920