Delete leftover AdServices files using a one-time boot-completed receiver
AdServices code is included in the ExtServices APK on S-, and generates multiple files that are stored in the data directory during normal operation. After OTA to T, the AdService code is removed from the ExtServices APK, but these data files are left behind. We're adding a new boot-completed receiver that's only enabled on T+ which will delete these files on reboot. After successful delete, the receiver disables itself since there's no need for it to run again.

Bug: 294912203
Test: atest, manual

Change-Id: I0fa1d452c4928cbc1c43efa5ba70124b75c12fd7