Fix AdServicesFilesCleanupBootCompleteReceiverHostTest
On T+ we are confirming the receiver is disabled by looking for a log statement output by the receiver. However, sometimes the process is killed before this can be logged, and the test fails. When this happens, we can also use package manager to check the status of the receiver.

Bug: 297712388
Test: presubmits, atest CtsExtServicesHostTests-tplus
Change-Id: I9b653d86b0344e895cd8dbd0ea449cb5576c3f85