[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
go/new-android-ownership-model

Bug: 304529413
Test: N/A
Change-Id: I3c4506b20eca7ff16456cf35cba17ca637c94f25
Merged-In: I7fbd55a92f4302a6e03bcff0283737fd9295165d