Fix failing test because assert was stronger than necessary
The test that verifies that the boot-complete receiver deletes adservices files within the ExtServices APK data directory was asserting that once the receiver finished, there would be no files present whose names contained the word "adservices". However, the code is only supposed to delete files whose names *begin with* the word "adservices". So, the assert was failing though the code was behaving correctly. Fixing the assert to match expected behavior.

Bug: 332178737
Test: atest CtsExtServicesHostTests-tplus

Change-Id: I266868c866349d3d1f41739ea93626c8de04f87b