Add force-no-test-error=false to libextservices_test
This option(newly added) tells GTest to not raise an exception if the test binary is missing.

Currently a missing binary is ignored silently by GTest. But we want to start throwing exceptions when it happens. We anticipated that there will be modules that are misconfigured already and so this option will be used to exempt such modules.

It seems that the module in this change are not configured properly and end up raising exception(WAI) when we enable the feature. so we are exempting them manually from the error. If you have a POC for these module, feel free to share with me and I will create a bug for them to fix it.

This is the context- b/352603943 and https://b.corp.google.com/issues/348446766#comment16

TLDR: The test modules are scheduled to run for both arm64 and arm abi, but they are not compiled for both abi. So the test module fails for the abi that was not compiled.

Merged-In: Ib9cdfac833df0b47effaa4c2251d7b6330ae6ca2
Bug: 352603943
Change-Id: I9253b4dd6c1773ae58b0d128c1eb052accd040a5
Test: presubmit