Exclude media and call style, refine end regex
Media and Call style notifications don't contain OTPs, and so should be
excluded from our OTP checker. Also ensures year false postives must be
the whole otp match (i.e. "11998" should not be a false positive just
because it includes "1998"). Finally, if the otp starts or ends with a
space, that space should not be preceded/followed by a number (so "43
G-3434 should not match), and the list of valid ending chars has been
shrunk

Bug: 350887669
Flag: NONE bugfix
Test: atest NotificationOtpDetectionHelperTest
Change-Id: I3072f20292aa32b1cc68f01a310d1c83ead4f89c